locations controller things
